+++ apps/api/src/operations/getNodeLineage.ts
@@ -8,6 +8,7 @@ import {
     NodeLineageParameters,
     NodeLinks,
     NODE_EMBEDDED_PARAMETERS,
+    TitledLink,
 } from "@phylopic/api-models"
 import { normalizeUUID, UUID } from "@phylopic/utils"
 import { ClientBase } from "pg"
@@ -26,21 +27,23 @@ import validate from "../validation/validate"
 import { Operation } from "./Operation"
 export type GetNodesParameters = DataRequestHeaders & NodeLineageParameters
 export type GetNodesService = PgClientService
+const DEFAULT_TITLE = "[Unnamed]"
 const ITEMS_PER_PAGE = 48
 const USER_MESSAGE = "There was a problem with a request to list taxonomic groups."
-const getQueryBuilder = (uuid: UUID, results: "total" | "uuid" | "json") => {
+const getQueryBuilder = (uuid: UUID, results: "total" | "href" | "json") => {
     const builder = new QueryConfigBuilder()
-    const selection = results === "total" ? 'COUNT("uuid") as total' : results === "uuid" ? '"uuid"' : '"json","uuid"'
+    const selection =
+        results === "total" ? 'COUNT("uuid") as total' : results === "href" ? 'title,"uuid"' : '"json","uuid"'
     const additional = results === "json" ? '"json",' : ""
     const additionalN = results === "json" ? 'n."json",' : ""
     builder.add(
         `
 WITH RECURSIVE predecessors AS (
-    SELECT ${additional}"uuid",parent_uuid,build,0 as lineage_index
+    SELECT ${additional}title,"uuid",parent_uuid,build,0 as lineage_index
         FROM node
         WHERE "uuid"=$::uuid AND build=$::bigint
     UNION
-    SELECT ${additionalN}n."uuid",n.parent_uuid,n.build,suc.lineage_index + 1
+    SELECT ${additionalN}n.title,n."uuid",n.parent_uuid,n.build,suc.lineage_index + 1
         FROM node n
         INNER JOIN predecessors suc ON suc.parent_uuid=n."uuid" AND suc.build=n.build
 )
@@ -49,7 +52,7 @@ SELECT ${selection} FROM predecessors
         [uuid, BUILD],
     )
     if (results !== "total") {
-        builder.add(`GROUP BY ${additional}"uuid",lineage_index ORDER BY lineage_index`)
+        builder.add(`GROUP BY ${additional}title,"uuid",lineage_index ORDER BY lineage_index`)
     }
     return builder
 }
@@ -60,11 +63,14 @@ const getTotalItems = (uuid: UUID) => async (client: ClientBase) => {
 }
 const getItemLinks =
     (uuid: UUID) =>
-    async (client: ClientBase, offset: number, limit: number): Promise<readonly Link[]> => {
-        const queryBuilder = getQueryBuilder(uuid, "uuid")
+    async (client: ClientBase, offset: number, limit: number): Promise<readonly TitledLink[]> => {
+        const queryBuilder = getQueryBuilder(uuid, "href")
         queryBuilder.add("OFFSET $ LIMIT $", [offset, limit])
-        const queryResult = await client.query<{ uuid: UUID }>(queryBuilder.build())
-        return queryResult.rows.map(({ uuid }) => ({ href: `/nodes/${uuid}?build=${BUILD}` }))
+        const queryResult = await client.query<{ title: string | null; uuid: UUID }>(queryBuilder.build())
+        return queryResult.rows.map(({ title, uuid }) => ({
+            href: `/nodes/${uuid}?build=${BUILD}`,
+            title: title || DEFAULT_TITLE,
+        }))
     }
 const getItemLinksAndJSON =
     (uuid: UUID) =>
@@ -73,17 +79,20 @@ const getItemLinksAndJSON =
         offset: number,
         limit: number,
         embeds: ReadonlyArray<string & keyof NodeEmbedded>,
-    ): Promise<ReadonlyArray<Readonly<[Link, string]>>> => {
+    ): Promise<ReadonlyArray<Readonly<[TitledLink, string]>>> => {
         const queryBuilder = getQueryBuilder(uuid, "json")
         queryBuilder.add("OFFSET $ LIMIT $", [offset, limit])
-        const queryResult = await client.query<{ json: string; uuid: UUID }>(queryBuilder.build())
+        const queryResult = await client.query<{ json: string; title: string | null; uuid: UUID }>(queryBuilder.build())
         if (!embeds.length) {
-            return queryResult.rows.map(({ json, uuid }) => [{ href: `/nodes/${uuid}?build=${BUILD}` }, json])
+            return queryResult.rows.map(({ json, title, uuid }) => [
+                { href: `/nodes/${uuid}?build=${BUILD}`, title: title || DEFAULT_TITLE },
+                json,
+            ])
         }
         return await Promise.all(
-            queryResult.rows.map(async ({ json, uuid }) => {
+            queryResult.rows.map(async ({ json, title, uuid }) => {
                 return [
-                    { href: `/nodes/${uuid}?build=${BUILD}` },
+                    { href: `/nodes/${uuid}?build=${BUILD}`, title: title || DEFAULT_TITLE },
                     await parseEntityJSONAndEmbed<Node, NodeLinks>(client, json, embeds, isNode, "taxonomic group"),
                 ]
             }),