+++ apps/api/src/pagination/getPageObject.ts
@@ -1,17 +1,18 @@
-import { Link, Page } from "@phylopic/api-models"
+import { Page, TitledLink } from "@phylopic/api-models"
 import { createSearch } from "@phylopic/utils"
+import omit from "lodash/omit"
 import BUILD from "../build/BUILD"
 const getPageObject = (
     listEndpoint: string,
     listQuery: Readonly<Record<string, string | number | boolean | undefined>>,
     index: number,
     lastPage: boolean,
-    items: readonly Link[],
+    items: readonly TitledLink[],
 ): Page => {
     return {
         _links: {
             items,
-            list: { href: listEndpoint + createSearch(listQuery) },
+            list: { href: listEndpoint + createSearch(omit(listQuery, "page")) },
             next: lastPage ? null : { href: listEndpoint + createSearch({ ...listQuery, page: index + 1 }) },
             previous: index > 0 ? { href: listEndpoint + createSearch({ ...listQuery, page: index - 1 }) } : null,
             self: { href: listEndpoint + createSearch({ ...listQuery, page: index }) },