+++ apps/www/src/views/ImageFilesView/index.tsx
@@ -1,5 +1,5 @@
-import { ImageWithEmbedded } from "@phylopic/api-models"
-import { useLicenseText, useNomenText } from "@phylopic/ui"
+import { Image } from "@phylopic/api-models"
+import { useLicenseText } from "@phylopic/ui"
 import { isString } from "@phylopic/utils"
 import { FC, useMemo } from "react"
 import slugify from "slugify"
@@ -8,22 +8,21 @@ import getImageFileExtension from "../../files/getImageFileExtension"
 import DownLoadLink from "./DownloadLink"
 import styles from "./index.module.scss"
 export interface Props {
-    value: ImageWithEmbedded
+    value: Image
 }
 const EXTENSION_LINKS: Readonly<Record<string, string>> = {
     png: "http://www.libpng.org/pub/png/spec/1.2/PNG-Contents.html",
     svg: "https://www.w3.org/TR/SVG/",
 }
 const ImageFilesView: FC<Props> = ({ value }) => {
-    const specificNameShort = useNomenText(value._embedded.specificNode?.names[0], true, "incertae sedis")
     const licenseShort = useLicenseText(value._links.license.href, true)
     const filenamePrefix = useMemo(
         () =>
-            [specificNameShort, value.attribution, licenseShort, value.uuid]
+            [value._links.self.title, value.attribution, licenseShort, value.uuid]
                 .filter(isString)
-                .map(s => slugify(s))
+                .map(s => slugify(s ?? ""))
                 .join("_"),
-        [value, licenseShort, specificNameShort],
+        [value, licenseShort],
     )
     const sourceFileExtension = useMemo(
         () => getImageFileExtension(value._links.sourceFile.type).toUpperCase(),