+++ packages/api-models/src/detection/isList.ts
@@ -5,7 +5,7 @@ import isData from "./isData"
 import isLink from "./isLink"
 import isLinks from "./isLinks"
 const isListLinks = (x: unknown, faultCollector?: ValidationFaultCollector): x is List["_links"] =>
-    isLinks(x) &&
+    isLinks(x, isLink(isNormalizedText), faultCollector) &&
     isNullOr(isLink(isNormalizedText))((x as List["_links"]).firstPage, faultCollector?.sub("firstPage")) &&
     isNullOr(isLink(isNormalizedText))((x as List["_links"]).lastPage, faultCollector?.sub("lastPage"))
 export const isList = (x: unknown, faultCollector?: ValidationFaultCollector): x is List =>