+++ packages/api-models/src/types/Image.ts
@@ -3,15 +3,16 @@ import { Entity } from "./Entity"
 import { Link } from "./Link"
 import { Links } from "./Links"
 import { MediaLink } from "./MediaLink"
-export interface ImageLinks extends Links {
-    readonly contributor: Link
-    readonly generalNode: Link | null
+import { TitledLink } from "./TitledLink"
+export interface ImageLinks extends Links<TitledLink> {
+    readonly contributor: TitledLink
+    readonly generalNode: TitledLink | null
     readonly "http://ogp.me/ns#image": MediaLink<URL, RasterMediaType>
     readonly license: Link<LicenseURL>
-    readonly nodes: readonly Link[]
+    readonly nodes: readonly TitledLink[]
     readonly rasterFiles: readonly MediaLink<URL, RasterMediaType>[]
     readonly sourceFile: MediaLink<URL>
-    readonly specificNode: Link
+    readonly specificNode: TitledLink
     readonly thumbnailFiles: readonly MediaLink<URL, RasterMediaType>[]
     // :TODO: Remove this line
     readonly "twitter:image"?: MediaLink<URL, RasterMediaType>