+++ apps/api-docs/CHANGELOG.md
@@ -19,6 +19,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.2.0] - 2023-03-21
+
+### Changed
+
+-   Many links changed to `TitledLink`.
+
+### Fixed
+
+-   Erroneous description of `Image.contributor.href`.
+
 ## [2.1.2] - 2023-03-10
 
 ### Changed