+++ apps/api/CHANGELOG.md
@@ -19,4 +19,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.2.0] - 2023-03-21
+
+### Changed
+
+-   Many links changed to `TitledLink`.
+
+### Fixed
+
+-   Invalid `list` link in page response.
+
 ## [2.1.1] - 2023-02-12