+++ apps/www/CHANGELOG.md
@@ -21,6 +21,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## [2.1.0] - 2023-03-21
+
+### Changed
+
+-   Added slugs to Contributor, Image, and Node Page URLs.
+-   Simplified silhouette `alt` descriptions.
+
 ## [2.0.10] - 2023-03-19
 
 ### Changed